--- src/main/java/kitchenpos/eatinorders/domain/order/EatInOrderCompletePolicy.java
@@ -1,20 +1,0 @@
-package kitchenpos.eatinorders.domain.order;
-
-import org.springframework.stereotype.Component;
-
-import kitchenpos.eatinorders.ordertable.domain.OrderTable;
-
-@Component
-public class EatInOrderCompletePolicy {
-    private final EatInOrderRepository eatInOrderRepository;
-
-    public EatInOrderCompletePolicy(EatInOrderRepository eatInOrderRepository) {
-        this.eatInOrderRepository = eatInOrderRepository;
-    }
-
-    public void follow(OrderTable orderTable) {
-        if (!eatInOrderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED)) {
-            orderTable.clean();
-        }
-    }
-}
