--- src/main/java/kitchenpos/eatinorders/domain/order/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorders.domain.order;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, COMPLETED
-}
