+++ src/main/java/kitchenpos/eatinorders/order/domain/EatInOrderFactory.java
@@ -1,0 +1,51 @@
+package kitchenpos.eatinorders.order.domain;
+
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Component;
+
+import kitchenpos.eatinorders.order.domain.vo.MenuVo;
+import kitchenpos.eatinorders.order.domain.vo.OrderLineItemVo;
+import kitchenpos.eatinorders.ordertable.domain.OrderTable;
+import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
+import kitchenpos.menus.domain.Price;
+
+@Component
+public class EatInOrderFactory {
+    private final OrderTableRepository orderTableRepository;
+    private final MenuApiRepository menuApiRepository;
+
+    public EatInOrderFactory(
+            final OrderTableRepository orderTableRepository,
+            final MenuApiRepository menuApiRepository) {
+        this.orderTableRepository = orderTableRepository;
+        this.menuApiRepository = menuApiRepository;
+    }
+
+    public EatInOrder create(UUID orderTableId, List<OrderLineItemVo> orderLineItemVos) {
+        List<OrderLineItem> orderLineItems = orderLineItemVos.stream()
+                .map(this::createOrderLineItem)
+                .collect(Collectors.toList());
+
+        final OrderTable orderTable = orderTableRepository.findById(orderTableId)
+                .orElseThrow(NoSuchElementException::new);
+
+        return new EatInOrder(UUID.randomUUID(), new OrderLineItems(orderLineItems), orderTable);
+    }
+
+    private OrderLineItem createOrderLineItem(OrderLineItemVo orderLineItemVo) {
+        final MenuVo menuVo = menuApiRepository.findById(orderLineItemVo.getMenuId());
+        if (!menuVo.isDisplayed()) {
+            throw new IllegalStateException();
+        }
+        Price menuPrice = new Price(menuVo.getPrice());
+        Price orderLineItemPrice = new Price(orderLineItemVo.getPrice());
+        if (menuPrice.isNotSame(orderLineItemPrice)) {
+            throw new IllegalArgumentException();
+        }
+        return new OrderLineItem(orderLineItemVo.getMenuId(), orderLineItemVo.getQuantity(), orderLineItemVo.getPrice());
+    }
+}
