+++ src/main/java/kitchenpos/eatinorders/order/domain/MenuApiRepository.java
@@ -1,0 +1,9 @@
+package kitchenpos.eatinorders.order.domain;
+
+import java.util.UUID;
+
+import kitchenpos.eatinorders.order.domain.vo.MenuVo;
+
+public interface MenuApiRepository {
+    MenuVo findById(UUID menuId);
+}
