+++ src/main/java/kitchenpos/eatinorders/order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorders.order.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
