+++ src/main/java/kitchenpos/eatinorders/order/infra/JpaEatInOrderRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.eatinorders.order.infra;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.eatinorders.order.domain.EatInOrder;
+import kitchenpos.eatinorders.order.domain.EatInOrderRepository;
+
+public interface JpaEatInOrderRepository extends EatInOrderRepository, JpaRepository<EatInOrder, UUID> {
+}
