--- src/main/java/kitchenpos/eatinorders/ui/EatInOrderRestController.java
@@ -1,55 +1,0 @@
-package kitchenpos.eatinorders.ui;
-
-import java.net.URI;
-import java.util.List;
-import java.util.UUID;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.eatinorders.application.EatInOrderService;
-import kitchenpos.eatinorders.application.dto.CreateOrderRequest;
-import kitchenpos.eatinorders.domain.order.EatInOrder;
-
-@RequestMapping("/api/orders")
-@RestController
-public class EatInOrderRestController {
-    private final EatInOrderService eatInOrderService;
-
-    public EatInOrderRestController(final EatInOrderService eatInOrderService) {
-        this.eatInOrderService = eatInOrderService;
-    }
-
-    @PostMapping
-    public ResponseEntity<EatInOrder> create(@RequestBody final CreateOrderRequest request) {
-        final EatInOrder response = eatInOrderService.create(request);
-        return ResponseEntity.created(URI.create("/api/orders/" + response.getId()))
-                .body(response);
-    }
-
-    @PutMapping("/{orderId}/accept")
-    public ResponseEntity<EatInOrder> accept(@PathVariable final UUID orderId) {
-        return ResponseEntity.ok(eatInOrderService.accept(orderId));
-    }
-
-    @PutMapping("/{orderId}/serve")
-    public ResponseEntity<EatInOrder> serve(@PathVariable final UUID orderId) {
-        return ResponseEntity.ok(eatInOrderService.serve(orderId));
-    }
-
-    @PutMapping("/{orderId}/complete")
-    public ResponseEntity<EatInOrder> complete(@PathVariable final UUID orderId) {
-        return ResponseEntity.ok(eatInOrderService.complete(orderId));
-    }
-
-    @GetMapping
-    public ResponseEntity<List<EatInOrder>> findAll() {
-        return ResponseEntity.ok(eatInOrderService.findAll());
-    }
-}
