+++ src/test/java/kitchenpos/eatinorders/order/support/InMemoryEatInOrderRepository.java
@@ -1,0 +1,40 @@
+package kitchenpos.eatinorders.order.support;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
+
+import kitchenpos.eatinorders.order.domain.EatInOrder;
+import kitchenpos.eatinorders.order.domain.EatInOrderRepository;
+import kitchenpos.eatinorders.order.domain.OrderStatus;
+import kitchenpos.eatinorders.ordertable.domain.OrderTable;
+
+public class InMemoryEatInOrderRepository implements EatInOrderRepository {
+    private final Map<UUID, EatInOrder> orders = new HashMap<>();
+
+    @Override
+    public EatInOrder save(final EatInOrder order) {
+        orders.put(order.getId(), order);
+        return order;
+    }
+
+    @Override
+    public Optional<EatInOrder> findById(final UUID id) {
+        return Optional.ofNullable(orders.get(id));
+    }
+
+    @Override
+    public List<EatInOrder> findAll() {
+        return new ArrayList<>(orders.values());
+    }
+
+    @Override
+    public boolean existsByOrderTableAndStatusNot(final OrderTable orderTable, final OrderStatus status) {
+        return orders.values()
+                .stream()
+                .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
+    }
+}
