+++ src/test/java/kitchenpos/eatinorders/ordertable/service/OrderTableServiceTest.java
@@ -15,7 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import kitchenpos.eatinorders.ordertable.application.OrderTableService;
 import kitchenpos.eatinorders.ordertable.application.dto.ChangeNumberOfGuestsRequest;
 import kitchenpos.eatinorders.ordertable.application.dto.CreateOrderTableRequest;
-import kitchenpos.eatinorders.domain.order.EatInOrderRepository;
+import kitchenpos.eatinorders.order.domain.EatInOrderRepository;
 import kitchenpos.eatinorders.ordertable.domain.OrderTable;
 import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
